HV-1667 Simplify ValidatorFactoryImpl
